Working on script to build a complete distribution package including all language bindings

+++ build.sh
@@ -1,25 +1,56 @@
-JAVA_HOME=/usr/lib/jvm/java-7-openjdk-amd64 # Ununtu 12.04 64-bit
-TIGHTDB_JAVA_HOME="$HOME/tightdb_java2"
+cd "$(dirname "$0")"
+TIGHTDB_JAVA_HOME="$(pwd)"
+
+MAKE="make -j8"
+
+# Setup OS specific stuff
+OS="$(uname -o)" || exit 1
+if [ "$OS" = "Darwin" ]; then
+    MAKE="$MAKE CC=gcc"
+fi
+
+# Find 'jni.h', 'java' and 'javac'
+if [ -z "$JAVA_HOME" ]; then
+    if ! which javac >/dev/null; then
+        echo "No JAVA_HOME and no Java compiler in PATH" 1>&2
+        exit 1
+    fi
+    JAVAC=$(readlink -f "$(which javac)")
+    if ! echo "$JAVAC" | grep '/bin/javac$' >/dev/null 2>&1; then
+        echo "Could not determine JAVA_HOME from path of 'javac' command" 1>&2
+        exit 1
+    fi
+    JAVA_HOME="$(echo "$JAVAC" | sed 's|/bin/javac$||')"
+fi
+for x in include/jni.h bin/java bin/javac; do
+    if ! [ -f "$JAVA_HOME/$x" ]; then
+        echo "No '$x' in '$JAVA_HOME'" 1>&2
+        exit 1
+    fi
+done
+JAVA="$JAVA_HOME/bin/java"
+JAVAC="$JAVA_HOME/bin/javac"
+JAVA_VERSION="$(java -version 2>&1 | grep '^java version' | sed 's/.*"\(.*\).*"/\1/')"
+JAVA_MAJOR="$(echo "$JAVA_VERSION" | cut -d. -f1)"
+JAVA_MINOR="$(echo "$JAVA_VERSION" | cut -d. -f2)"
+if [ "$JAVA_MAJOR" -lt 1 -o \( "$JAVA_MAJOR" -eq 1 -a "$JAVA_MINOR" -lt 6 \) ]; then
+    echo "Need Java version 1.6 or newer (is '$JAVA_VERSION')" 1>&2
+    exit 1
+fi
+
 
 
 # Build libtightdb-jni.so
 cd "$TIGHTDB_JAVA_HOME/tightdb_jni/src" || exit 1
-make clean || exit 1
-make -j8 EXTRA_CFLAGS="-I\"$JAVA_HOME/include\"" || exit 1
+$MAKE clean || exit 1
+$MAKE EXTRA_CFLAGS="-I\"$JAVA_HOME/include\"" || exit 1
 
 # Build tightdb.jar
 cd "$TIGHTDB_JAVA_HOME/src/main/java" || exit 1
 find com/ -type f -name '*.class' -delete || exit 1
 export CLASSPATH=/usr/share/java/commons-io.jar:/usr/share/java/commons-lang.jar:/usr/share/java/freemarker.jar:.
-javac $(find com/ -type f -name '*.java' | fgrep -v /doc/ | fgrep -v /example/) || exit 1
+$JAVAC $(find com/ -type f -name '*.java' | fgrep -v /doc/ | fgrep -v /example/) || exit 1
 jar cf tightdb.jar $(find com/ -type f -name '*.class') || exit 1
-
-# Build and run example
-cd "$TIGHTDB_JAVA_HOME/tightdb-example/src/main/java" || exit 1
-find com/ -type f -name '*.class' -delete || exit 1
-export CLASSPATH="$TIGHTDB_JAVA_HOME/src/main/java/tightdb.jar:."
-javac com/tightdb/example/Example.java com/tightdb/example/generated/*.java || exit 1
-java -Djava.library.path="$TIGHTDB_JAVA_HOME/tightdb_jni/src" com.tightdb.example.Example || exit 1
 
 # Build and run test suite
 cd "$TIGHTDB_JAVA_HOME/src/test/java" || exit 1
@@ -27,5 +58,5 @@ find com/ -type f -name '*.class' -delete || exit 1
 test_sources="$(find * -type f -name '*Test.java')"
 test_classes="$(echo "$test_sources" | sed 's/\.java$/.class/')"
 export CLASSPATH="$TIGHTDB_JAVA_HOME/src/main/java/tightdb.jar:$TIGHTDB_JAVA_HOME/tightdb-example/src/main/java:/usr/share/java/testng.jar:/usr/share/java/qdox.jar:/usr/share/java/bsh.jar:."
-javac $test_sources || exit 1
-java -Djava.library.path="$TIGHTDB_JAVA_HOME/tightdb_jni/src" org.testng.TestNG -d "$TIGHTDB_JAVA_HOME/test_output" -testclass $test_classes || exit 1
+$JAVAC $test_sources || exit 1
+$JAVA -Djava.library.path="$TIGHTDB_JAVA_HOME/tightdb_jni/src" org.testng.TestNG -d "$TIGHTDB_JAVA_HOME/test_output" -testclass $test_classes || exit 1
